fix: format confetti
